Add files via upload
Encompasses a systematic approach to conducting sentiment analysis on 
Twitter data. It begins with data collection and preprocessing to clean and standardize the 
text, followed by exploratory analysis to gain insights into the dataset. Machine learning 
algorithms are trained and evaluated to predict tweet sentiments accurately, with optimization techniques applied to improve performance. Finally, the optimized model is deployed for real-time sentiment analysis, enabling its use in the domain for recognize the racist tweets.